+++ components/vf-discussion/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.11 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.10 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.9 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.8 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.7 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.11 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.10 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.9 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.8 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.7 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
+## 0.0.6 (2019-06-03)
+
+**Note:** Version bump only for package @visual-framework/vf-discussion
+
+
+
+
+
 ## 0.0.5 (2019-05-02)
 
 **Note:** Version bump only for package @visual-framework/vf-discussion