+++ components/vf-grid/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.36 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.35 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.34 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.33 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.32 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.36 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.35 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.34 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.33 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.32 (2019-06-20)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
+## 0.0.31 (2019-06-03)
+
+**Note:** Version bump only for package @visual-framework/vf-grid
+
+
+
+
+
 ## 0.0.29 (2019-05-02)
 
 **Note:** Version bump only for package @visual-framework/vf-grid